Moved game reset processing into main game framework. Also modified game start processing on the server so that we don't want for the IN_PLAY attribute change to be processed before calling gameDidStart() because we don't want to do that when resetting the game (because someone could come along and end the game in between gameWillStart() and gameDidStart() which would be annoying and bad), so we need to be consistent and run straight through both when starting the game for the first time and when resetting it.
This means that game managers that really need to wait until their events
are processed and do some game start processing at that point will have to
listen for the IN_PLAY transition themselves *and* they'll have to be
aware that someone could have come along and ended the game in between the
call to gameDidStart() and the dispatch of the IN_PLAY event. Sketchy, but
unavoidable.


git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@1235 542714f4-19e9-0310-aa3c-eee0fc999fb1