Refactored things so that the whirled.spot Location class is merely an interface and specific games can define their own concept of a location. Portal and SceneLocation now both have a Location instance within them. All the whirled stuff is happy not knowing the details of the Location even though it manages portals and other such goodies. The 'stage' stuff does need to know, and yohoho uses the new StageLocation class that is essentially identical to the old 'Location'.
SceneLocation previously extended Location, and there were a lot of places
where they were used somewhat interchangeably, or a SceneLocation and
a regular Location were compared. This shouldn't be done anymore and I've
left in code that should dump stack if it occurs, so that we can spot
these logic errors at runtime.


git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@4140 542714f4-19e9-0310-aa3c-eee0fc999fb1