Fixup. If there's an error when calling the callback function, we were assuming it was an argument-related error and trying again. However, this is sorta bad behavior as we don't want to call twice into a callback. In the case I was seeing, the first time in caused an NPE, but popped up a panel, and the second try was no-opping because the panel was already up (but misconfigured).
So: don't even try a second call, because it looks like
flash will accept args to an argless function.
We'll have to try this out for a while and see if it works.
If so, I can neaten this code. If not, I'll uncomment
some of this and try a more surgical approach (only catch
ArgumentError, only retry if we had 1 boolean arg.)


git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@5585 542714f4-19e9-0310-aa3c-eee0fc999fb1